fix spec references, remove blank files
