+++ test/resources/TextCorpus.spec.ts
@@ -1,19 +1,15 @@
-
-process.env.NODE_ENV = 'test';
-
 import * as nock from 'nock';
 import * as Chance from 'chance';
-import {suite, test} from 'mocha-typescript';
+import { suite, test } from 'mocha-typescript';
 import * as sinon from 'sinon';
 import * as chai from 'chai';
 import * as sinonChai from 'sinon-chai';
-
-import {Factory, HudAiClientConfiguration} from '../../lib/util/ClientConfigFactory';
-import {TokenManager} from '../../lib/TokenManager';
-import {Session} from '../../lib/Session';
-import {RequestManager} from '../../lib/RequestManager';
 import * as moment from 'moment';
-import {TextCorpusResource} from '../../lib/resources/TextCorpus';
+
+import { Factory } from '../../lib/util/ClientConfigFactory';
+import { HudAiClient, HudAiClientConfiguration } from '../../lib/HudAiClient';
+import { RequestManager } from '../../lib/RequestManager';
+import { TextCorpusResource } from '../../lib/resources/TextCorpus';
 
 chai.use(sinonChai);
 nock.disableNetConnect();
@@ -22,21 +18,20 @@ const chance = new Chance();
 const clientId = chance.guid();
 const clientSecret = chance.guid();
 const accessToken = chance.guid();
+const refreshToken = chance.guid();
 
 @suite
-class TextCorpusSpec {
+class ArticleSpec {
     private sandbox: any;
     private config: HudAiClientConfiguration;
-    private tokenManager: TokenManager;
-    private session: Session;
     private requestManager: RequestManager;
+    private client: HudAiClient;
 
     before() {
         this.sandbox = sinon.sandbox.create();
         this.config = Factory({ clientId, clientSecret });
-        this.requestManager = new RequestManager(this.config);
-        this.tokenManager = new TokenManager(this.config, this.requestManager);
-        this.session = new Session(accessToken, this.tokenManager);
+        this.client = new HudAiClient(this.config);
+        this.requestManager = new RequestManager(this.client, this.config);
     }
 
     after() {
@@ -45,7 +40,8 @@ class TextCorpusSpec {
 
     @test
     instantiate() {
-        const resource = new TextCorpusResource(this.session, this.requestManager);
+        const config = Factory({ clientId: chance.guid(), clientSecret: chance.guid() });
+        const resource = new TextCorpusResource(this.requestManager);
         expect(resource.get).to.be.a('function');
         expect(resource.list).to.be.a('function');
         expect(resource.create).to.be.a('function');
